merge to master.

+++ app/src/main/java/com/darjeelingteagarden/model/MyOrder.kt
@@ -4,6 +4,6 @@ data class MyOrder(
     val orderId: String,
     val orderDate: String,
     val totalItems: Int,
-    val totalPrice: Int,
+    val totalPrice: Double,
     val currentStatus: String
 )
